move files according to new folders
